move tests to module all
This removes an ambiguity about how the tests are started up by cargo.
Cargo presumes any *.rs file is its own crate, and fires the tests
accordingly.  This means tests in the top level of the tests/ directory
cannot share any common test functions, like almost_eq.

I don't want almost_eq to be redefined in each crate, so it's declared
in integration_tests.rs.  For tests to use this function
definition, they have to be in the integration_tests crate.  Hence why
this commit moves them into all, which is a module in the
integration_tests crate.